refactor: move proxy response template components in api package
