+++ gravitee-apim-console-webui/src/management/api/response-templates/edit/api-response-templates-edit.component.html
