+++ gravitee-apim-console-webui/src/management/api/response-templates/edit/api-response-templates-edit.component.scss
@@ -1,8 +1,8 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
-@use '../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-content-container;
