+++ gravitee-apim-console-webui/src/management/api/response-templates/list/api-response-templates-list.component.html
