+++ gravitee-apim-console-webui/src/management/api/response-templates/list/api-response-templates-list.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-content-container;
