+++ gravitee-apim-console-webui/src/management/api/response-templates/list/api-response-templates-list.component.ts
@@ -20,18 +20,18 @@ import { EMPTY, of, Subject } from 'rxjs';
 import { catchError, filter, switchMap, takeUntil, tap } from 'rxjs/operators';
 import { ActivatedRoute } from '@angular/router';
 
-import { SnackBarService } from '../../../../../services-ngx/snack-bar.service';
-import { GioPermissionService } from '../../../../../shared/components/gio-permission/gio-permission.service';
+import { SnackBarService } from '../../../../services-ngx/snack-bar.service';
+import { GioPermissionService } from '../../../../shared/components/gio-permission/gio-permission.service';
 import { ResponseTemplate, toResponseTemplates } from '../response-templates.adapter';
-import { onlyApiV1V2Filter, onlyApiV2Filter } from '../../../../../util/apiFilter.operator';
-import { ApiV2Service } from '../../../../../services-ngx/api-v2.service';
+import { onlyApiV1V2Filter, onlyApiV2Filter } from '../../../../util/apiFilter.operator';
+import { ApiV2Service } from '../../../../services-ngx/api-v2.service';
 
 @Component({
   selector: 'api-proxy-response-templates-list',
-  templateUrl: './api-proxy-response-templates-list.component.html',
-  styleUrls: ['./api-proxy-response-templates-list.component.scss'],
+  templateUrl: './api-response-templates-list.component.html',
+  styleUrls: ['./api-response-templates-list.component.scss'],
 })
-export class ApiProxyResponseTemplatesListComponent implements OnInit, OnDestroy {
+export class ApiResponseTemplatesListComponent implements OnInit, OnDestroy {
   private unsubscribe$: Subject<boolean> = new Subject<boolean>();
 
   public responseTemplateTableDisplayedColumns = ['key', 'contentType', 'statusCode', 'actions'];
