+++ gravitee-apim-console-webui/src/management/api/response-templates/response-templates.adapter.ts
@@ -15,7 +15,7 @@
  */
 import { flatMap } from 'lodash';
 
-import { ApiV1, ApiV2 } from '../../../../entities/management-api-v2';
+import { ApiV1, ApiV2 } from '../../../entities/management-api-v2';
 
 export type ResponseTemplate = {
   id: string;
